:x
This update ensures the mc interface is 8GHz wider than
the nmc interface.

APIImpact

This example will setup a service with a mc width of 37.5GHz and
a nmc width of 29.5GHz. Note the property "service-format": "other".

If the property "service-format" is set to e.g. "OTU", TPCE will
setup a service where the nmc interface width is 40GHz and
the mc interface width is 37.5GHz. Setting the service-format
to 'other' resolves the issue.

POST /rests/operations/org-openroadm-service:service-create
{
    "input": {
        ...
        "service-a-end": {
            ...
            "service-format": "other",
            "frequency-slot": {
                "center-frequency": 193.1,
                "slot-width": 37.5
            },
            ...
        },
        "service-z-end": {
            ...
            "service-format": "other",
            "frequency-slot": {
                "center-frequency": 193.1,
                "slot-width": 37.5
            },
            ...
        }
    }
}

JIRA: TRNSPRTPCE-833
Change-Id: I82a130c0cac5647caf91f22c53825b3620e5c001
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>