Implement center-freq-granularity in freq. select
When a frequency range is selected for a new service,
center-freq-granularity on the nodes is taken into account.

Since each SRG/DEG may have different center-freq-granularity values,
the least common multiplier is used to find a "common ground".

This commit also uses the "base frequency" (i.e. 193.1) as the
reference point when finding center frequencies in a BitSet
representing available frequencies (as opposed to the start/end of
the BitSet).

Previously, the range was selected by walking "backwards" in the
frequency range (high to low). Now the iteration direction is
configurable using the environment variable
PCE_SPECTRUM_LOCATION_STRATEGY. The variable accepts on of two values:

    - HIGH_TO_LOW_FREQUENCY (default)
    - LOW_TO_HIGH_FREQUENCY

Not setting the variable will cause TPCE to use the backwards compatible
approach.

EXAMPLE

    export PCE_SPECTRUM_LOCATION_STRATEGY=LOW_TO_HIGH_FREQUENCY

JIRA: TRNSPRTPCE-849
Change-Id: I0f36d4868d40423290c51cf83a2bf22c1bf21021
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>