Implement center-freq-granularity in freq. select
When a frequency range is selected for a new service,
center-freq-granularity on the nodes are taken into account.

Since each SRG/DEG may have different center-freq-granularity values,
the least common multiplier is used to find a "common ground".

This commit also uses the "base frequency" 193.1 (G.694.1) as the
reference point, when finding possible center frequencies, as opposed
to the start/end of the range.

In short, this commit implements the leaf center-freq-granularity
defined in org-openroadm-device:

    leaf center-freq-granularity {
      type org-openroadm-common-optical-channel-types:frequency-GHz;
      default "50";
      config false;
      description
        "Granularity of allowed center frequencies.
         The base frequency for this computation is 193.1 THz (G.694.1)"
    }

Example #1
----------

    _____                      _____
   |     | 37.5GHz      75GHz |     |
   |  A  |____________________|  B  |
   |_____|                    |_____|

Given the above scenario, and assuming the base frequency for this
computation is 193.1 THz (G.694.1), 193.025 and 193.175 (THz) are
possible center frequencies compatible with both A and B.
In this case, valid center frequencies are 193.1 ± n x 75GHz and
will be used when searching for a valid service slot.

   A,B            A            A,B            A            A,B
    ↓             ↓             ↓             ↓             ↓
... |-------------|-------------|-------------|-------------| ...
 193.025       193.0625      193.1         193.1375      193.175

     <- 37.5GHz -> <- 37.5GHz -> <- 37.5GHz -> <- 37.5GHz ->
     <-------   75GHz   -------> <-------   75GHz   ------->

Searching for a valid frequency slot
------------------------------------

By default, the range is selected by walking "backwards" in the
frequency range (high to low). The iteration direction is now also
configurable using the environment variable:
PCE_SPECTRUM_LOCATION_STRATEGY.

The variable accepts one of two values:

    - HIGH_TO_LOW_FREQUENCY (default)
    - LOW_TO_HIGH_FREQUENCY

Not setting the variable will cause TPCE to use the backwards compatible
approach (i.e. HIGH_TO_LOW_FREQUENCY).

EXAMPLE

    export PCE_SPECTRUM_LOCATION_STRATEGY=LOW_TO_HIGH_FREQUENCY

JIRA: TRNSPRTPCE-849
Change-Id: I0f36d4868d40423290c51cf83a2bf22c1bf21021
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>