Migrate last remaining uses of InstanceIdentifier
In the TPCE code, calls to the toLegacy() method are hanging around
since the migration from InstanceIdentifier to DataObjectIdentifier,
because they were needed to use the firstKeyOf() method. Now that the
latter is available in DataObjectIdentifier, we can now remove calls to
toLegacy.
Also add few SuppressWarnings annotations to remove few unchecked casts.

JIRA: TRNSPRTPCE-842
Change-Id: I280aea51025779a51981506d759f85def2f52f28
Signed-off-by: Gilles Thouenon <gilles.thouenon@orange.com>