Implement center-freq-granularity in freq. select
When a frequency range is selected for a new service,
center-freq-granularity on the nodes are taken into account.

Since each node may have different center-freq-granularity values,
the least common multiplier is used to find a "common ground".

This commit also uses the "base frequency" 193.1 (G.694.1) as the
reference point, when finding possible center frequencies, as opposed
to the start/end of the range.

In short, this commit implements the leaf center-freq-granularity
defined in org-openroadm-device:

    leaf center-freq-granularity {
      type org-openroadm-common-optical-channel-types:frequency-GHz;
      default "50";
      config false;
      description
        "Granularity of allowed center frequencies.
         The base frequency for this computation is 193.1 THz (G.694.1)"
    }

Example
-------

In this example two nodes have these center-freq-granularity values:

    _____                      _____
   |     | 37.5GHz      75GHz |     |
   |  A  |____________________|  B  |
   |_____|                    |_____|

The center frequency granularity 75GHz is compatible with both
A and B above.

Given the scenario above, potential center frequencies are
193.1THz ± n x 75GHz, where n≥0 and will be used while searching for
an available service slot. To be more exact, this range will
be used: 191.325THz < 193.1THz ± n x 75GHz < 196.125THz.

             A,B         A         A,B         A         A,B
              ↓          ↓          ↓          ↓          ↓
  ...-|--...--|----------|----------|----------|----------|------|
   193.1   195.95     195.9875   196.025    196.0625   196.1  196.125

       <-...-> <- 37.5 -> <- 37.5 -> <- 37.5 -> <- 37.5 -> < 25 >
       <-...-> <------ 75.0  ------> <------ 75.0  ------>

JIRA: TRNSPRTPCE-849
Change-Id: I0f36d4868d40423290c51cf83a2bf22c1bf21021
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>