Adding Mockito java agent and explicit version
In order for Mockito to work correctly in the future it needs
to be added to the jvm explicitly in runtime when the tests are
run. This can be done with the surefire plugin but it needs the
path to the Mockito library, which contains the version.

In order to get the version it needs to be stated explicitly
somewhere. I chose to put this in the POM at module level,
because not all modules will use this, and also, modules
could benifit from choosing their own explicit version or
testing library as well.

Tl;DR;
* Added explicit versions of the surefire maven plugin and
  mockito-core to POM fiels using mockito in the tests.
* Added the maven surefire plugin and configured it accordingly.

Change-Id: Ifefb55bb30c3460a32b5279e522a55359dacbbd6
Signed-off-by: Jimmi Wimmersjö <jimmi.wimmersjo@smartoptics.com>