Added plant uml diagrams
Added plant uml diagrams for the listererns in ServiceHandler
and TAPI modules. This to easier get an overview of the message
system within TPCE.

Also included a (hopefully thorough) diagram for ServiceCreate

Also made 2 "happy path" (lightweight) diagrams for ServiceCreate
and ServiceDelete.

Summerized the appearant intent, and actions of the listener
classes in the "notifications.md" file.

Change-Id: Ib2e45d43f3022a61817a5501af4246dd54525302
Signed-off-by: Jimmi Wimmersjö <jimmi.wimmersjo@smartoptics.com>