Added plant uml documentation
Added a happy path (lightweight, only inludes notifications) for
ServiceCreate and ServiceDelete.

The reason for adding this documentation is to simplify the
notification architecture so it is easier for new people to
understand and easier for developers and architects to reason and
discuss the message and work flows.

Added rendered pictures of the UML charts into the docs/images folder.

Added a section in the developer guide with regards to the listeners
where these pictures are linked.

Added plantuml charts in docs/plantuml.

Change-Id: I08ed209b51f652ba7dbc88ead4d814320ad32dee
Signed-off-by: Jimmi Wimmersjö <jimmi.wimmersjo@smartoptics.com>