Added available frequencies when occupied frequencies exists
For some reason, when the tapi initialization encountered an SRG
and there where some occupied frequencies, the available frequencyMap
was never populated.

This code now populates the available frequency map with the complement,
to used frequencies, or it can be views as
supportedFrequencies - usedFrequencies intervalls.

Added a unit test for the createAvailableFreqMap method.

Change-Id: I2b0184ff0995f5d18381febad103ce1df04decb0
Signed-off-by: Jimmi Wimmersjö <jimmi.wimmersjo@smartoptics.com>