Added available frequencies
For some reason, when the tapi initialization encountered an SRG
and there where some occupied frequencies, the available frequencyMap
was never populated.

This code now populates the available frequency map with upperbound,
lowerBound IF the usedFrequencyMap is null or empty.

If the usedFrequencyMap has a frequency it returns an empty map.

Added unit tests for the respctive two cases:
usedFreqMap is empty
UsedFreqMap is not empty.

Jira: TRNSPRTPCE-869
Change-Id: I2b0184ff0995f5d18381febad103ce1df04decb0
Signed-off-by: Jimmi Wimmersjö <jimmi.wimmersjo@smartoptics.com>