Remove String dependency in TAPI ConnectivityUtils
CHANGES

ConnectivityUtils uses the openroadm topology when determining
a termination point and node type.

PROBLEMS ADDRESSED IN THIS COMMIT

Previously, the code relied on specific strings being present in
termination point IDs to determine the type.

This method required a ROADM node id to match a specific pattern,
e.g. transforming "ROADM-A1-SRG1" to "ROADM-A1":

  private String getIdBasedOnModelVersion(String nodeid) {
    return nodeid.matches("[A-Z]{5}-[A-Z0-9]{2}-.*")
      ? String.join("-", nodeid.split("-")[0], nodeid.split("-")[1])
      : nodeid.substring(0, nodeid.lastIndexOf("-"));
  }

Additionally, the following code relied on hard-coded strings in the
termination point IDs.
Now the topology is used to determine the types.

  if (tpID.contains("CLIENT")) {
    ...
  }
  if (tpID.contains("NETWORK")) {
    ...
  }
  if (tpID.contains("PP")) {
    tpNode = getIdBasedOnModelVersion(tp.getTpNodeId());
    ...
  }
  if (tpID.contains("TTP")) {
    tpNode = getIdBasedOnModelVersion(tp.getTpNodeId());
    ...
  }
  if (nodeId.contains("XPDR") || nodeId.contains("SPDR")
      || nodeId.contains("MXPDR")) {
    ...
  }
  if (nodeId.contains("ROADM")) {
    nodeId = getIdBasedOnModelVersion(nodeId);
    ...
  }

JIRA: TRNSPRTPCE-889
Change-Id: Idab55b8183c4392ddbc4ef9ea552cfd33fed320f
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>