Prevent exception during PCE spectrum allocation
This commit will prevent a IndexOutOfBoundsException during spectrum
allocation. The exception may occurr if the available spectrum is less
than the requested bandwidth and the avaialable spectrum is located at
the lowest frequencies.

EXAMPLE

Requested bandwidth is 8 bits (e.g. 8 x 6.25GHz), and the available
bandwidth is 4 bits (4 x 6.25GHz), located according to the picture
below, where 1 is available spectrum and 0 is occupied.

   191.325 THz
      |--------|---...
       11110000 000...

In this scenario a IndexOutOfBoundsException was thrown. This commit
mitigates the problem.

JIRA: TRNSPRTPCE-894
Change-Id: I74efc95db547060113b3a7f08ba568f4e3b61753
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>