Refactor PostAlgoPathValidatorTest replace XStream
The unit tests in PostAlgoPathValidatorTest that verify the method
getSpectrumAssignment in PostAlgoPathValidator, are refactored.

This commit replaces the dependency on XStream for reading XML files
into Java objects with Mockito-based mock objects.

The change reduces complexity, removes the dependency on XStream, and
makes PostAlgoPathValidatorTest easier to maintain.

CHANGES

The method getSpectrumAssignment in the PostAlgoPathValidator class
accepts four parameters, of which the following two are relevant to this
change:
    - GraphPath<String, PceGraphEdge> path
    - Map<NodeId, PceNode> allPceNodes

The tests in PostAlgoPathValidatorTest have been refactored to create
GraphPath<String, PceGraphEdge> mock objects.

PostAlgoPathValidatorTest now uses the new MockPceNodeMapFactory class,
which provides a method for creating Map<NodeId, PceNode> mock objects.

PREVIOUSLY

PostAlgoPathValidatorTest used XStream to read XML files located in:
  - resources/topology/links
    These files were used to create GraphPath<String, PceGraphEdge>
    objects.
  - resources/topology/nodes
    These files were used to create Map<NodeId, PceNode> objects.

Change-Id: I859243df68886aa44fe08d4881f61b9e0ce6d3f9
Signed-off-by: Joakim Törnqvist <joakim.tornqvist@smartoptics.com>